setMove()、actMove()、Objectを現在の位置から指定した位置に動かす関数を実装した。ただし、移動するスピードやフレーム間の時間などを考慮した調整が必要。 Buttonクラスの実装に取り掛かる。

+++ Title.h
@@ -11,7 +11,7 @@ private:
 	Text* titleText;
 	SDL_Color color;
 
-	int count = 0;
+	float count = 0.0f;
 
 	vector<Object*> drawList;
 public:
